+++ src/index.tsx
@@ -5,7 +5,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './react_apps/App';
+import App from './components/App';
 
 ReactDOM.render(
   <React.StrictMode>
